+++ test/creates/send_rcs.test.js
@@ -1,0 +1,17 @@
+const zapier = require('zapier-platform-core');
+
+const App = require('../../index');
+const appTester = zapier.createAppTester(App);
+zapier.tools.env.inject();
+
+describe('creates.send_rcs', () => {
+  it('should run', async () => {
+    const bundle = { inputData: {} };
+
+    const results = await appTester(
+      App.creates['send_rcs'].operation.perform,
+      bundle,
+    );
+    expect(results).toBeDefined();
+  });
+});